Address an issue caused by empty input when uploading
If a user only selected an image and left all other fields blank
when uploading a recipe, nulls (expected empty strings) would be
passed to server and a recipe with null properties would then be
stored in database and cause error on search, as to the parse
function could not handle null input. Now the upload function
checks null in requests, and the parse function also handles
null input.